Viele Änderungen

+++ database/migrations/2008_05_28_094549_create_one_time_payments_table.php
@@ -20,8 +20,8 @@ class CreateOneTimePaymentsTable extends Migration
             $table->unsignedBigInteger('favoredFundsCenter_id');
             $table->unsignedBigInteger('chargedFundsCenter_id');
             $table->unsignedBigInteger('costType_id');
-            $table->smallInteger('grantedFunds');
-            $table->smallInteger('spentFunds')->nullable();
+            $table->integer('grantedFunds');
+            $table->integer('spentFunds')->nullable();
             $table->date('spentDate')->nullable();
             $table->string('notes', 500)->nullable();
             $table->string('requirements', 500)->nullable();
